refactor: reorganize module directories
