+++ docs/source/modules/index.rst
@@ -4,7 +4,16 @@ Modules
 Database
 --------
 
-.. automodule:: json_as_db.Database
+.. automodule:: json_as_db.core.database
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+Matcher
+--------
+
+.. automodule:: json_as_db.core.matcher
    :members:
    :undoc-members:
    :show-inheritance:
