+++ src/json_as_db/constants.py
